requirements.txt
ultralytics>=8.1.0
opencv-python>=4.8.0
pyyaml>=6.0.1
tqdm>=4.66.0
numpy>=1.24.0
streamlit>=1.30.0
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
